feat(rust-axum): DBまで通る結合テスト追加・本番バグ修正 #19
- 18本の統合テスト追加（auth/clients/staffs/invitation/gate/notifications）
- Axum 0.8 ルートパラメータ構文を :param → {param} に修正
- SQLx 0.8 の INT UNSIGNED 型厳密検査に対応（i32→u32）
- DBパスワードの # 文字を URL エンコードして DSN 解析バグを修正
- tests/common/mod.rs にテスト用スキーマ作成・ヘルパー関数を実装
- CI（rust-ci.yml）に統合テスト実行ステップを追加

+++ backends/rust-axum/src/infrastructure/persistence/client.rs
@@ -28,7 +28,7 @@ struct ClientRow {
     private_key:  String,
     public_key:   String,
     fingerprint:  String,
-    status:       i32,
+    status:       u32,
     start_at:     Option<NaiveDateTime>,
     stop_at:      Option<NaiveDateTime>,
     created_at:   NaiveDateTime,
@@ -37,7 +37,7 @@ struct ClientRow {
     updated_by:   Option<u32>,
     deleted_at:   Option<NaiveDateTime>,
     deleted_by:   Option<u32>,
-    version:      i32,
+    version:      u32,
 }
 
 fn row_to_entity(r: ClientRow) -> Client {
@@ -56,7 +56,7 @@ fn row_to_entity(r: ClientRow) -> Client {
         private_key:  r.private_key,
         public_key:   r.public_key,
         fingerprint:  r.fingerprint,
-        status:       r.status,
+        status:       r.status as i32,
         start_at:     r.start_at,
         stop_at:      r.stop_at,
         created_at:   r.created_at,
@@ -65,7 +65,7 @@ fn row_to_entity(r: ClientRow) -> Client {
         updated_by:   r.updated_by,
         deleted_at:   r.deleted_at,
         deleted_by:   r.deleted_by,
-        version:      r.version,
+        version:      r.version as i32,
     }
 }
